requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
asyncpg==0.29.0
Babel==2.13.1
boto3==1.33.10
botocore==1.33.10
certifi==2023.11.17
cffi==1.16.0
cfgv==3.4.0
click==8.1.7
cryptography==42.0.0
distlib==0.3.7
dnspython==2.5.0
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.104.1
fastapi-pagination==0.12.13
fastapi-storages==0.2.1
filelock==3.13.1
graphql-core==3.2.3
greenlet==3.0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.26.0
Hypercorn==0.15.0
hyperframe==6.0.1
identify==2.5.32
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
jmespath==1.0.1
Mako==1.3.0
MarkupSafe==2.1.3
nodeenv==1.8.0
packaging==23.2
Pillow==10.1.0
platformdirs==4.0.0
pluggy==1.3.0
pre-commit==3.5.0
priority==2.0.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
pytest==7.4.3
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
s3transfer==0.8.2
setuptools==69.0.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sse-starlette==1.8.2
starlette==0.27.0
strawberry-graphql==0.217.1
typing_extensions==4.8.0
urllib3==2.0.7
virtualenv==20.24.7
wsproto==1.2.0
